requirements.txt
openai>=1.0.0
python-dotenv
langchain>=0.0.1
langchain-openai>=0.2.0
langchain-core>=0.3.0
